Move BaseCard.vue to utils
