+++ frontend/components/auth/FormLogin.vue
@@ -43,7 +43,7 @@
 <script lang="ts">
 import Vue from 'vue'
 import { userNameRules, passwordRules } from '@/rules/index'
-import BaseCard from '@/components/molecules/BaseCard.vue'
+import BaseCard from '@/components/utils/BaseCard.vue'
 
 export default Vue.extend({
   components: {
