+++ frontend/components/label/FormUpload.vue
@@ -38,7 +38,7 @@
 
 <script lang="ts">
 import Vue from 'vue'
-import BaseCard from '@/components/molecules/BaseCard.vue'
+import BaseCard from '@/components/utils/BaseCard.vue'
 import { uploadSingleFileRules } from '@/rules/index'
 
 export default Vue.extend({
