+++ frontend/components/member/FormCreate.vue
@@ -57,7 +57,7 @@
 
 <script lang="ts">
 import Vue, { PropType } from 'vue'
-import BaseCard from '@/components/molecules/BaseCard.vue'
+import BaseCard from '@/components/utils/BaseCard.vue'
 import { UserDTO } from '~/services/application/user.service'
 import { RoleDTO } from '~/services/application/role.service'
 import { MemberDTO } from '~/services/application/member.service'
