+++ frontend/components/organisms/annotation/GuidelineCard.vue
@@ -16,7 +16,7 @@
 import 'tui-editor/dist/tui-editor-contents.css'
 import 'highlight.js/styles/github.css'
 import { Viewer } from '@toast-ui/vue-editor'
-import BaseCard from '@/components/molecules/BaseCard'
+import BaseCard from '@/components/utils/BaseCard'
 import '@/assets/style/editor.css'
 
 export default {
