+++ frontend/components/project/FormCreate.vue
@@ -59,7 +59,7 @@
 
 <script lang="ts">
 import Vue from 'vue'
-import BaseCard from '@/components/molecules/BaseCard.vue'
+import BaseCard from '@/components/utils/BaseCard.vue'
 import { projectNameRules, descriptionRules, projectTypeRules } from '@/rules/index'
 
 export default Vue.extend({
