+++ frontend/components/project/FormDelete.vue
@@ -31,7 +31,7 @@
 
 <script lang="ts">
 import Vue, { PropType } from 'vue'
-import BaseCard from '@/components/molecules/BaseCard.vue'
+import BaseCard from '@/components/utils/BaseCard.vue'
 import { ProjectDTO } from '@/services/application/project.service'
 
 export default Vue.extend({
