+++ frontend/components/utils/BaseCard.vue
@@ -35,8 +35,9 @@
   </v-card>
 </template>
 
-<script>
-export default {
+<script lang="ts">
+import Vue from 'vue'
+export default Vue.extend({
   props: {
     title: {
       type: String,
@@ -65,5 +66,5 @@ export default {
       this.$emit('cancel')
     }
   }
-}
+})
 </script>
