+++ frontend/components/utils/ConfirmForm.vue
@@ -21,7 +21,7 @@
 
 <script lang="ts">
 import Vue from 'vue'
-import BaseCard from '@/components/molecules/BaseCard.vue'
+import BaseCard from './BaseCard.vue'
 
 export default Vue.extend({
   components: {
